Update program 06.py

+++ ProgramBatch2/program 06.py
@@ -1,2 +1,2 @@
-numbers = [int(input(f"Enter number {i+1}: ")) for i in range(10)]
+numbers = [float(input(f"Enter number {i+1}: ")) for i in range(10)]
 print("Result:", numbers[0] - sum(numbers[1:]))
